fix: 视频组增删改

+++ business/src/main/java/com/abdecd/moebackend/business/service/backstage/impl/VideoGroupServiceImpl.java
@@ -1,8 +1,9 @@
 package com.abdecd.moebackend.business.service.backstage.impl;
 
 import com.abdecd.moebackend.business.common.exception.BaseException;
-import com.abdecd.moebackend.business.common.util.SpringContextUtil;
-import com.abdecd.moebackend.business.dao.entity.*;
+import com.abdecd.moebackend.business.dao.entity.PlainUserDetail;
+import com.abdecd.moebackend.business.dao.entity.Video;
+import com.abdecd.moebackend.business.dao.entity.VideoGroup;
 import com.abdecd.moebackend.business.dao.mapper.*;
 import com.abdecd.moebackend.business.pojo.dto.backstage.bangumiVideoGroup.BangumiVideoGroupUpdateDTO;
 import com.abdecd.moebackend.business.pojo.dto.backstage.commonVideoGroup.VideoGroupDTO;
@@ -11,28 +12,21 @@ import com.abdecd.moebackend.business.pojo.vo.backstage.commonVideoGroup.VideoGr
 import com.abdecd.moebackend.business.pojo.vo.backstage.commonVideoGroup.VideoVo;
 import com.abdecd.moebackend.business.pojo.vo.plainuser.UploaderVO;
 import com.abdecd.moebackend.business.pojo.vo.statistic.StatisticDataVO;
-import com.abdecd.moebackend.business.pojo.vo.videogroup.PlainVideoGroupVO;
 import com.abdecd.moebackend.business.service.ElasticSearchService;
-import com.abdecd.moebackend.business.service.fileservice.FileService;
 import com.abdecd.moebackend.business.service.backstage.VideoGroupService;
-import com.abdecd.moebackend.business.service.plainuser.PlainUserService;
+import com.abdecd.moebackend.business.service.fileservice.FileService;
 import com.abdecd.moebackend.business.service.statistic.StatisticService;
 import com.abdecd.moebackend.business.service.video.VideoService;
 import com.abdecd.moebackend.business.service.videogroup.BangumiVideoGroupServiceBase;
 import com.abdecd.moebackend.business.service.videogroup.PlainVideoGroupServiceBase;
 import com.abdecd.moebackend.business.service.videogroup.VideoGroupServiceBase;
-import com.abdecd.moebackend.common.constant.MessageConstant;
-import com.abdecd.moebackend.common.constant.RedisConstant;
 import com.abdecd.moebackend.common.constant.VideoGroupConstant;
 import com.abdecd.tokenlogin.common.context.UserContext;
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import jakarta.annotation.Resource;
 import jakarta.validation.Valid;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.tomcat.util.buf.StringUtils;
 import org.springframework.beans.BeanUtils;
-import org.springframework.cache.annotation.CacheEvict;
-import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
@@ -40,7 +34,6 @@ import org.springframework.web.multipart.MultipartFile;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Objects;
-import java.util.stream.Collectors;
 
 @Service
 @Slf4j
@@ -58,13 +51,7 @@ public class VideoGroupServiceImpl implements VideoGroupService {
     private StatisticService statisticService;
 
     @Resource
-    private VideoGroupTagMapper videoGroupTagMapper;
-
-    @Resource
     private PlainUserDetailMapper plainUserDetailMapper;
-
-    @Resource
-    private VideoGroupAndTagMapper videoGroupandTagMapper;
 
     @Resource
     private VideoService videoService;
@@ -154,8 +141,13 @@ public class VideoGroupServiceImpl implements VideoGroupService {
         videoGroup.setCover(coverPath);
 
         videoGroupMapper.update(videoGroup);
+        var newOne = videoGroupServiceBase.getVideoGroupInfo(videoGroup.getId());
 
-        elasticSearchService.deleteSearchEntity(videoGroup.getId());
+        if (newOne != null) {
+            elasticSearchService.saveSearchEntity(newOne);
+        } else {
+            elasticSearchService.deleteSearchEntity(videoGroup.getId());
+        }
     }
 
     @Override
@@ -275,17 +267,21 @@ public class VideoGroupServiceImpl implements VideoGroupService {
             }
         }
 
-        videoGroupMapper.update(
-                new VideoGroup()
-                        .setId(videoGroup.getId())
-                        .setVideoGroupStatus(Byte.valueOf(videoGroup.getVideoGroupStatus()))
-                        .setTitle(videoGroup.getTitle())
-                        .setCover(coverPath)
-                        .setDescription(videoGroup.getDescription())
-                        .setTags(videoGroup.getTags())
-        );
+        var entity = new VideoGroup()
+                .setId(videoGroup.getId())
+                .setVideoGroupStatus(Byte.valueOf(videoGroup.getVideoGroupStatus()))
+                .setTitle(videoGroup.getTitle())
+                .setCover(coverPath)
+                .setDescription(videoGroup.getDescription())
+                .setTags(videoGroup.getTags());
+        videoGroupMapper.update(entity);
 
-        elasticSearchService.deleteSearchEntity(videoGroup.getId());
+        var newOne = videoGroupServiceBase.getVideoGroupInfo(entity.getId());
+        if (newOne != null) {
+            elasticSearchService.saveSearchEntity(newOne);
+        } else {
+            elasticSearchService.deleteSearchEntity(entity.getId());
+        }
     }
 
     @Override
